Move datastore tests into a separate module
The sheer number of tests in the datastore module required this to
improve readability of the code.
